-changed price from float to int -reimplemented UpdateBuyerPriceBelief, UpdateSellerPriceBelief

+++ Assets/Scripts/AuctionRecord.cs
@@ -5,12 +5,12 @@
     public int Consumed { get; internal set; }
     public int Bids { get; internal set; }
     public int Offers { get; internal set; }
-    public float AvgOfferPrice { get; internal set; }
-    public float AvgBidPrice { get; internal set; }
-    public float AvgClearingPrice { get; internal set; }
-    public float Stocks { get; internal set; }
-    public float Capitals { get; internal set; }
-    public float Profits { get; internal set; }
-    public float Cashs { get; internal set; }
+    public int AvgOfferPrice { get; internal set; }
+    public int AvgBidPrice { get; internal set; }
+    public int AvgClearingPrice { get; internal set; }
+    public int Stocks { get; internal set; }
+    public int Capitals { get; internal set; }
+    public int Profits { get; internal set; }
+    public int Cashs { get; internal set; }
     public int Professions { get; internal set; }
 }
