+++ Assets/Scripts/TradeRecord.cs
@@ -6,7 +6,7 @@ public class TradeRecord
 {
     public string ItemName { get; internal set; }
     public TransactionType TransactionType { get; internal set; }
-    public float Price { get; internal set; }
+    public int Price { get; internal set; }
     public int Quantity { get; internal set; }
     public int Round { get; internal set; }
 }
